Move BlockType implementations to their own package
