+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/world/blocks/types/impl/BlockTypeGrass.java
@@ -1,4 +1,6 @@
-package io.github.willqi.pizzaserver.server.world.blocks.types;
+package io.github.willqi.pizzaserver.server.world.blocks.types.impl;
+
+import io.github.willqi.pizzaserver.server.world.blocks.types.BlockTypeID;
 
 public class BlockTypeGrass extends BlockTypeFullSolid {
