+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/world/blocks/types/impl/BlockTypeStone.java
@@ -1,4 +1,4 @@
-package io.github.willqi.pizzaserver.server.world.blocks.types;
+package io.github.willqi.pizzaserver.server.world.blocks.types.impl;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
@@ -7,6 +7,7 @@ import io.github.willqi.pizzaserver.commons.utils.Vector3;
 import io.github.willqi.pizzaserver.nbt.tags.NBTCompound;
 import io.github.willqi.pizzaserver.nbt.tags.NBTString;
 import io.github.willqi.pizzaserver.server.item.ItemToolType;
+import io.github.willqi.pizzaserver.server.world.blocks.types.BlockTypeID;
 
 import java.util.*;
